+++ src/components/AllyRequestedCalls/AllyRequestedCalls.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { useHistory } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
 
-import RequestedCallItem from '../RequestedCallItem/RequestedCallItem';
+import RequestedCallItem from './Components/RequestedCallItem';
 
 function AllyRequestedCalls(){
